CodeTemplate/LunarLander_starter_code/requirements.txt
gymnasium==0.29.1
swig==4.2.1
matplotlib==3.5.3
ipython==7.34.0
moviepy==1.0.3
pyvirtualdisplay==3.0
opencv-python==4.6.0.66
ipdb==0.13.9
box2d-py==2.3.5
numpy==1.23.0
PyYAML==6.0.2
wandb==0.18.1
gymnasium[box2d]==0.29.1

